+++ src/app/frontend/chrome/chrome_controller.js
