+++ src/app/frontend/replicasetlist/replicasetlist_controller.js
