+++ src/test/integration/zerostate_test.js
